fixed encryption test suite
omg the intricate little bugs took me so long to find and fix and stuff.... i think the main problem was that the variable 'key' was being scoped to the parameters in an anonymous function in the .then,,, I must say it is rather confusing when it works perfectly in developer tools but not in code. Also I wrote a little custom encoder to just use some pastable characters because that also lead to problems.. yeah...
the code is now really ugly and messy and i understood it before but now i dont really but im too scared to touch it for now. At least it works (i think) :p